+++ packages/core/src/component.ts
@@ -16,7 +16,7 @@ import {
   ComponentProps,
 } from './component.types';
 
-export { uiElement, uiElements, uiEvent, MQBasedRendered, prop, tag } from './decorators';
+export { uiElement, uiElements, uiEvent, MQBasedRendered, prop, tag, renderAsync } from './decorators';
 
 export const INITIALIZED_EVENT = 'kl-component-initialized';
 
@@ -41,6 +41,8 @@ export class Component extends HTMLElement {
   protected reactions: ComponentReactions<this> = {
     initialized: ['onComponentInitialized'],
   };
+
+  protected _shouldRenderAsync?: boolean;
 
   private _state = {};
   private _initialStates = {};
@@ -86,6 +88,10 @@ export class Component extends HTMLElement {
 
     this.mergeEvents(events);
     this.props = this.normalizeProps({ ...(this[decoratedProps] || null), ...props });
+  }
+
+  private get asyncRenderingEnabled(): boolean {
+    return this._shouldRenderAsync || this.asyncRendering;
   }
 
   /* ====================================================
@@ -196,7 +202,7 @@ export class Component extends HTMLElement {
    * Description
    */
   private async setupComponent(): Promise<void> {
-    if (this.asyncRendering) {
+    if (this.asyncRenderingEnabled) {
       await this.renderAsync();
     } else {
       this.renderComponent();
@@ -252,7 +258,7 @@ export class Component extends HTMLElement {
       if (typeof this[event.handler] === 'function') {
         const targets = this.getEventTargets(event.target);
         // @ts-ignore
-        onEvent(targets, event.event, this[event.handler], this);
+        onEvent(targets, event.event, this[event.handler], this, event.options);
       }
     });
   }
@@ -265,7 +271,7 @@ export class Component extends HTMLElement {
       if (typeof this[event.handler] === 'function') {
         const targets = this.getEventTargets(event.target);
         // @ts-ignore
-        removeEvent(targets, event.event, this[event.handler], this);
+        removeEvent(targets, event.event, this[event.handler], this, event.options);
       }
     });
   }
@@ -288,9 +294,9 @@ export class Component extends HTMLElement {
       if (typeof this[event.handler] === 'function') {
         const targets = this.getEventTargets(event.target);
         // @ts-ignore
-        removeEvent(targets, event.event, this[event.handler], this);
+        removeEvent(targets, event.event, this[event.handler], this, event.options);
         // @ts-ignore
-        onEvent(targets, event.event, this[event.handler], this);
+        onEvent(targets, event.event, this[event.handler], this, event.options);
       }
     });
   }
@@ -308,12 +314,12 @@ export class Component extends HTMLElement {
       if (decoratorUiDef.selector === 'window') {
         decoratorUiDef.events.forEach((event) => {
           // @ts-ignore
-          onEvent(window, event.eventName, this[event.handler], this);
+          onEvent(window, event.eventName, this[event.handler], this, event.options);
         });
       } else if (decoratorUiDef.selector === 'this') {
         decoratorUiDef.events.forEach((event) => {
           // @ts-ignore
-          onEvent(this, event.eventName, this[event.handler], this);
+          onEvent(this, event.eventName, this[event.handler], this, event.options);
         });
       } else {
         const curEl = decoratorUiDef.justOne
@@ -323,7 +329,7 @@ export class Component extends HTMLElement {
         this[property] = curEl;
         decoratorUiDef.events.forEach((event) => {
           // @ts-ignore
-          onEvent(this[property], event.eventName, this[event.handler], this);
+          onEvent(this[property], event.eventName, this[event.handler], this, event.options);
         });
       }
     });
@@ -338,19 +344,19 @@ export class Component extends HTMLElement {
       if (decoratorUiDef.selector === 'window') {
         decoratorUiDef.events.forEach((event) => {
           // @ts-ignore
-          removeEvent(window, event.eventName, this[event.handler], this);
+          removeEvent(window, event.eventName, this[event.handler], this, event.options);
         });
       } else if (decoratorUiDef.selector === 'this') {
         decoratorUiDef.events.forEach((event) => {
           // @ts-ignore
-          removeEvent(this, event.eventName, this[event.handler], this);
+          removeEvent(this, event.eventName, this[event.handler], this, event.options);
         });
       } else {
         const curEl = this[property];
         if (curEl !== null && curEl !== undefined) {
           decoratorUiDef.events.forEach((event) => {
             // @ts-ignore
-            removeEvent(this[property], event.eventName, this[event.handler], this);
+            removeEvent(this[property], event.eventName, this[event.handler], this, event.options);
           });
         }
         // @ts-ignore