+++ packages/core/src/decorators/renderAsync.ts
@@ -1,0 +1,5 @@
+export default function renderAsync<T extends { new (...args: any[]): {} }>(constructor: T) {
+  return class extends constructor {
+    _shouldRenderAsync = true;
+  };
+}